backend/requirements.txt
fastapi
uvicorn[standard]
crewai
crewai[tools]
openai
python-dotenv
langchain-google-genai
langchain-community
databases>=0.5.5
aiosqlite